requirements.txt
# OCT Extractor Application Dependencies

# Core library for OCT file processing
oct-converter>=0.4.0

# GUI framework
PyQt5>=5.15.0

# Data manipulation and image processing
numpy>=1.19.0
Pillow>=8.0.0

# File handling and data storage
h5py>=3.1.0

# Visualization
matplotlib>=3.3.0
